+++ src/accounts/context/userAccount.tsx
@@ -3,12 +3,11 @@ import React, {FC, useCallback, useEffect, useState} from 'react'
 import {useDispatch} from 'react-redux'
 
 // Types
-import {Account as UserAccount, postCancel} from 'src/client/unityRoutes'
+import {Account as UserAccount} from 'src/client/unityRoutes'
 
 // Notifications
 import {notify} from 'src/shared/actions/notifications'
 import {
-  accountCancellationError,
   accountDefaultSettingError,
   accountDefaultSettingSuccess,
   accountRenameError,
@@ -25,8 +24,6 @@ import {
 // Metrics
 import {event} from 'src/cloud/utils/reporting'
 import {isFlagEnabled} from 'src/shared/utils/featureFlag'
-import {useHistory} from 'react-router-dom'
-import {getErrorMessage} from 'src/utils/api'
 
 // Actions
 import {setMe} from 'src/me/actions/creators'
@@ -37,7 +34,6 @@ export type Props = {
 }
 export interface UserAccountContextType {
   userAccounts: UserAccount[]
-  handleCancelAccount: () => void
   handleGetAccounts: () => void
   handleSetDefaultAccount: (newId: number) => void
   handleRenameActiveAccount: (accountId: number, newName: string) => void
@@ -49,7 +45,6 @@ export const DEFAULT_CONTEXT: UserAccountContextType = {
   userAccounts: [],
   defaultAccountId: -1,
   activeAccountId: -1,
-  handleCancelAccount: () => {},
   handleGetAccounts: () => {},
   handleSetDefaultAccount: () => {},
   handleRenameActiveAccount: () => {},
@@ -65,7 +60,6 @@ export const UserAccountProvider: FC<Props> = React.memo(({children}) => {
   const [activeAccountId, setActiveAccountId] = useState<number>(null)
 
   const dispatch = useDispatch()
-  const history = useHistory()
 
   /**
    * get the name of the account as specified by ID
@@ -80,20 +74,6 @@ export const UserAccountProvider: FC<Props> = React.memo(({children}) => {
       return selectedAcctArray[0].name
     }
   }
-
-  const handleCancelAccount = useCallback(async () => {
-    try {
-      const resp = await postCancel({})
-
-      if (resp.status !== 204) {
-        throw new Error(resp.data.message)
-      }
-    } catch (error) {
-      const message = getErrorMessage(error)
-      console.error({error})
-      dispatch(notify(accountCancellationError(message)))
-    }
-  }, [dispatch, history])
 
   const handleGetAccounts = useCallback(async () => {
     try {
@@ -181,7 +161,6 @@ export const UserAccountProvider: FC<Props> = React.memo(({children}) => {
         userAccounts,
         defaultAccountId,
         activeAccountId,
-        handleCancelAccount,
         handleGetAccounts,
         handleSetDefaultAccount,
         handleRenameActiveAccount,
